inital refactor for watch list separation
